raspberry/requirements.txt
certifi==2024.2.2
influxdb3-python==0.3.6
numpy==1.26.4
paho-mqtt==2.0.0
pyarrow==15.0.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
reactivex==4.0.4
six==1.16.0
typing_extensions==4.9.0
urllib3==2.2.1
